requirements.txt
Django-3.2.8
Pillow==8.3.2
python-dotenv==0.19.1
django-phonenumber-field==6.0.0
phonenumbers==8.12.38